Delete RedditMarkdown
